Changed user entity a bit and reworked register dto

+++ fitness-club/src/main/java/com/fitness_club/user/model/User.java
@@ -35,9 +35,9 @@ public class User {
 
     private int age;
 
-    private double height;
+    private int height;
 
-    private double weight;
+    private int weight;
 
     @Enumerated(EnumType.STRING)
     private Gender gender;
